# api/internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "milmil API",
    "version": "1.0.0",
    "description": "Self-hosted anime media server"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/api/v1/discover/anime/{id}": {
      "get": {
        "summary": "Anime detail by Bangumi ID, enriched with AniList cover and banner",
        "operationId": "getAnimeDetail",
        "tags": [
          "Discover"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Anime detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnimeDetail"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID"
          },
          "404": {
            "description": "Anime not found"
          },
          "429": {
            "description": "Upstream rate limited"
          },
          "502": {
            "description": "External service unavailable"
          }
        }
      }
    },
...
    "/api/v1/discover/browse": {
      "get": {
        "summary": "Browse anime by genre",
        "operationId": "browseByGenre",
        "tags": [
          "Discover"
        ],
        "parameters": [
          {
            "name": "genre",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Anime results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnimeSummary"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/discover/browse/tag": {
      "get": {
        "summary": "Browse anime by tag",
        "operationId": "browseByTag",
        "tags": [
          "Discover"
        ],
        "parameters": [
          {
            "name": "tag",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Anime results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnimeSummary"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/discover/calendar": {
      "get": {
        "summary": "Daily airing schedule from Bangumi.tv",
        "operationId": "getCalendar",
        "tags": [
          "Discover"
        ],
        "responses": {
          "200": {
            "description": "Array of calendar days with anime per weekday",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CalendarDay"
                  }
                }
              }
            }
          },
          "429": {
            "description": "Upstream rate limited"
          },
          "502": {
            "description": "External service unavailable"
          }
        }
      }
    },
...
    "/api/v1/discover/trending": {
      "get": {
        "summary": "Global trending anime from AniList, enriched with Bangumi Chinese titles",
        "operationId": "getTrending",
        "tags": [
          "Discover"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of trending anime",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnimeSummary"
                  }
                }
              }
            }
          },
          "429": {
            "description": "Upstream rate limited"
          },
          "502": {
            "description": "External service unavailable"
          }
        }
      }
    },
...
    "/api/v1/libraries/{id}": {
      "get": {
        "summary": "Get a library by ID",
        "operationId": "getLibrary",
        "tags": [
          "Libraries"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Library details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Library"
                }
              }
            }
          },
          "404": {
            "description": "Library not found"
          }
        }
      },
      "put": {
        "summary": "Update a library",
        "operationId": "updateLibrary",
        "tags": [
          "Libraries"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "path"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "path": {
                    "type": "string"
                  },
                  "enabled": {
                    "type": "boolean"
                  },
                  "scan_interval_minutes": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Library updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Library"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Library not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a library",
        "operationId": "deleteLibrary",
        "tags": [
          "Libraries"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Library deleted"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API token (mlml_ prefix)",
        "description": "Pass an API token obtained from POST /api/v1/auth/login or POST /api/v1/api-tokens. Tokens have the prefix `mlml_`."
      }
    },
    "schemas": {
      "AnimeDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/AnimeSummary"
          },
          {
            "type": "object",
            "properties": {
              "synopsis": {
                "type": "string",
                "description": "Chinese synopsis from Bangumi"
              },
              "banner_image": {
                "type": "string"
              },
              "tags": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "popularity": {
                "type": "integer"
              },
              "rating": {
                "$ref": "#/components/schemas/Rating"
              }
            }
          }
        ]
      },
      "AnimeSummary": {
        "type": "object",
        "properties": {
          "bangumi_id": {
            "type": "integer"
          },
          "anilist_id": {
            "type": "integer"
          },
          "title": {
            "type": "string",
            "description": "Chinese title preferred, fallback Japanese"
          },
          "title_original": {
            "type": "string",
            "description": "Original Japanese title"
          },
          "title_en": {
            "type": "string"
          },
          "cover_image": {
            "type": "string"
          },
          "air_date": {
            "type": "string"
          },
          "episode_count": {
            "type": "integer"
          },
          "score": {
            "type": "number"
          }
        }
      },
...
      "CalendarDay": {
        "type": "object",
        "properties": {
          "weekday": {
            "type": "string",
            "description": "Chinese weekday name"
          },
          "weekday_en": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AnimeSummary"
            }
          }
        }
      },
...
      "Library": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "enabled": {
            "type": "integer",
            "description": "1 = enabled, 0 = disabled"
          },
          "scan_interval_minutes": {
            "type": "integer"
          },
          "last_scanned_at": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        }
      },
...
      "Rating": {
        "type": "object",
        "properties": {
          "score": {
            "type": "number"
          },
          "total": {
            "type": "integer"
          }
        }
      },
...
    }
  },
...
}